TINYGL: Look ahead on divergence.
Do not systematically advance on both previous and current frame
call lists. Check a few draw calls ahead if a match is found, and stall
previous frame call list in that case. Useful when a few calls are inserted
in the middle of a frame (like Chinchila Charlie's smoke puffs in GRIM set
"cn", or text lines), saving the rest of the frame from irretrievably
diverging.