GRIM: Do not draw when flipping is forbidden.
Fixes
"Allocator out of memory: couldn't allocate more memory from linear allocator."
error with TinyGL in GRIM in several places:
- at the beginning of Year 2, after Meche boards the boat (does not happen
  when skipping cutscene)
- when loading set bi
And likely in other places.
The issue is that TinyGL queues draw calls in a fixed-size structure
(linear allocator used to quickly allocate memory to queued vertices),
and it emits that exception when running out of space.
When _flipEnable is false (controlled by game scripts), no drawing happens,
so the draw call queue is not cleared. The longer flipping is forbidden,
and the more complex the scene is (more vertices, like in bi), the more
likely it is to run out of space and cause residualvm to error out.
In set bi, missing one flip is enough to exceed buffer space.
In set lm (the one after meche boarding the boat), it takes 5 missed flips
to exceed buffer space (simpler scene), but it still happens in a split
second.
There is no way with OpenGL to discard already-submitted calls (which was
my first idea for TinyGL), so do not do this.
It seems likely OpenGL to have larger draw operation buffers (maybe
dynamically extended ?), and/or to forcibly stall caller and execute draw
operations when running out of space. In any case, this would add
complexity to residualvm and use more CPU and GPU time for no purpose.
So instead just skip drawing until flipping is allowed again.